(re task future) Use ES2015 classes for futures and tasks
